refactor(tui): drop dead code and unused options across components
Remove unreachable functions, unused functional options and
test-only helpers from the production code paths in pkg/tui.
No behavior change; builds, lint, and the full test suite all
pass.

- scrollview: remove unused WithGapWidth option.
- sidebar: remove unused WithLayoutConfig and Option type, drop
  the variadic opts ...Option from New (no caller passes any).
- reasoningblock: remove the BlockMsg/blockMsgBase/ToggleMsg/
  GetBlockID infrastructure; nothing constructed a ToggleMsg.
- messages: drop the now-unreachable BlockMsg switch case in
  Update and the orphaned forwardToReasoningBlock helper.
- tool/registry: collapse Register/RegisterAll into a single
  Register(registrations) entry point; the per-name overload
  was never called.
- markdown: move newGlamourRenderer into fast_renderer_test.go
  (only tests used it); production renderer.go no longer needs
  the glamour or styles imports.
- message: move stripANSI into message_test.go (only tests
  used it); production message.go no longer needs regexp.
